+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/buttons/reply/ReplyKeyboardButtonsShortcuts.kt
@@ -124,7 +124,7 @@ inline fun requestUsersOrBotsReplyButton(
 inline fun requestUserOrBotReplyButton(
     text: String,
     requestId: RequestId
-) = requestUserReplyButton(
+) = requestUsersReplyButton(
     text,
     KeyboardButtonRequestUsers.Any(requestId)
 )